fix(transport/grpc): remove autoprefixing logic
Removes the autoprefixing of "pb" to service names, to allow for
grpc services created without a package prefix.

Fixes #447.

BREAKING CHANGE: Existing users of the grpc client transport who were
relying on the autoprefixing of the "pb" package name must manually
update their clients to add it. No change is required if a package name
was already being manually prefixed.